Handle WebGL context creation error in `createWebGL2Context`
The function also sets the clear color and the viewport on success. Even
though it might not be an ideal design in a real app, here it makes sure
that the "framework" is easy to use and takes care of the irrelevant
details. The viewport must have been set everytime anyway, and the clear
color can be changed at any point.