Preserve aspect ratio when rendering
Instead of providing a whole perspective camera, I've decided to go with
a simpler route, and that is: just make sure we keep track of how much
bigger/smaller the width/height are, and provide that info to the
shader.

A camera is not needed anyway, since pretty much everything happens in
the fragment shader. Any "camera" work will be done there.